+++ src/ui/handler/scroll_zoom.ts
@@ -297,7 +297,7 @@ export class ScrollZoomHandler implements Handler {
             }
 
             const fromScale = typeof this._targetZoom !== 'number' ? tr.scale : zoomScale(this._targetZoom);
-            this._targetZoom = tr.getConstrained(tr.getCameraLngLat(), scaleZoom(fromScale * scale)).zoom;
+            this._targetZoom = tr.constrain(tr.getCameraLngLat(), scaleZoom(fromScale * scale)).zoom;
 
             // if this is a mouse wheel, refresh the starting zoom and easing
             // function we're using to smooth out the zooming between wheel